added read time machine from config
