+++ conf/application.conf
@@ -168,7 +168,6 @@ feature-switch {
   enable-mfa-credits-and-debits = true
   enable-payment-history-refunds = true
   enable-iv-uplift = true
-  enable-time-machine = false
   enable-income-sources = true
   enable-calendar-quarter-types = true
   enable-adjust-payments-on-account = true
@@ -304,6 +303,7 @@ auth {
 }
 
 time-machine {
+    enable-time-machine = false
     add-years = 1
     add-days = 0
 }
